Добавил еще один абстракт класс и нужно его протестить, изменил view.start, view.finish,try with resource
